Code Refactored - Move the loader into the GroceryItemLoader Class & Other Changes like constants file added
