+++ app/src/main/java/com/retail/retailapp/util/GroceryConstant.java
@@ -1,0 +1,20 @@
+package com.retail.retailapp.util;
+
+/**
+ * Created by raghuramankumarasamy on 20/10/16.
+ */
+public class GroceryConstant {
+    public static final String QUANTITY_MAP = "QuantityMap";
+    public static final String GROCERYITEMS_JSON = "groceryitems.json";
+    public static final String SL_NO = " Sl.No   ";
+    public static final String PRODUCT = " Product                     ";
+    public static final String UNIT = " Unit            ";
+    public static final String PRICE = " Price    ";
+
+    private GroceryConstant() {
+
+    }
+    public static final String GROCERY_MAP = "GroceryMap";
+
+
+}
